refactor(tools): move investigation-feedback to workflow/deep-search
It's a write operation (adjusts retrieval weighting), not a read-only
query. It directly depends on deep_search's investigationId and MemoryRef
— placing it in query/ created a cross-directory reverse dependency.
