+++ src/tools/workflow/deep-search/investigation-feedback.ts
@@ -1,7 +1,7 @@
 import { Type } from "@sinclair/typebox";
 import type { ToolDefinition } from "@mariozechner/pi-coding-agent";
-import { FEEDBACK_SIGNALS, type FeedbackStatus } from "../../memory/types.js";
-import type { MemoryRef } from "../workflow/deep-search/tool.js";
+import { FEEDBACK_SIGNALS, type FeedbackStatus } from "../../../memory/types.js";
+import type { MemoryRef } from "./tool.js";
 
 interface FeedbackParams {
   investigationId: string;
